feat(solver): index canonical declaration-site defs
